+++ challenges.html
@@ -46,7 +46,7 @@
         <nav class="footer-social">
             <h2>Follow us on</h2>
             <menu>
-                <li><a href="">Facebook</a></li><i class="fa-solid fa-star"></i>
+                <li><a href="">Facebook</a></li>
                 <li><a href="">Instagram</a></li>
                 <li><a href="">Twitter</a></li>
             </menu>
